Handle permissions.
Of course, this is a really simple way to handle the permissions. Any method
listed in the "readonly_method" will have the "readonly" permission required,
and all the other ones will require the "readwrite" permission.

One benenfit of defining this in the ViewSet class is that implementers can
shortcut any complexity here if they need, by always returning the same fixed
string.