Handle CORS settings.
Create a subclass of the default cornice "Service" class, rather than
cluttering the default ones with defaults that are sensible only for cliquet.

I though it would be possible to add the CORS information at registration time
(when the venusian callback is called again in the "register_resource"
function, but it would have worked only for resources.