Add request bound data, shared with subrequests.
This allows to store any arbitrary data in a ``dict`` on the current request,
instead of setting custom attributes and testing their existance.

It is shared with subrequests during batch.

In the short term, this will allow us to reduce storage hits in batch
subrequests on the default bucket in Kinto.

In the long term, this will allow us to share storage connections accross
batch subrequests.