Load plugins after setting up authn/authz
While developing kinto-attachment, we realized that if plugins were included
before setting-up authn/authz, their views could not be protected by permissions.

For example, when a plugin defines a plugin like this in a view:

```python

    upload = Service(name="upload",
                     path="/upload",
                     description="",
                     factory=RouteFactory)
    @upload.get(permission="upload")
    def upload_get(request):
        return {"ok": True}

```

The factory is taken into account, but the authorization policy was never
called!

With this change, the plugins are included after ``cliquet.initialization.setup_authentication()``. Thus after ``config.include("pyramid_multiauth")`` thus after setting
up the authorization policy from setting ``multiauth.authorization_policy``.

The side-effet of this is that plugins cannot set default settings regarding auth.
It was the case of *cliquet-fxa* which defines an implicit setting ``multiauth.policy.fxa.use`` that allowed to just to specify ``fxa`` in the ``multiauth.policies`` setting.

The setting ``multiauth.policy.fxa.use`` must now be defined explicitly.