rename get-started to install
This section doesn't really get anyone started. The real "get started"
documentation would be the "first steps" page that shows real API calls
and their responses.

This section used to be two sections, one of which was called
installation, and the other of which was called run-kinto, so reclaim
the "installation" name.