move tests from kinto/core to kinto/tests/core
Anything in kinto/ that isn't in kinto/tests/ has to have 100%
coverage. That doesn't seem valuable to me.

At the same time, we want to maintain a distinction between the "core"
part of the library which people might build servers out of, and the
non-core thing that people will just run.

kinto/tests/core seems like a good compromise.