get rid of kinto/core/scripts/cliquet
There's really only one thing here that's valuable, and that's the
actual migrate command. Leave that around in case someone else wants
it. The fact that kinto, the major cliquet application to date, doesn't
use anything else here, suggests we don't need it. Perhaps we were
keeping it around for the sake of having an example for other
authors. In that case, the kinto script is probably sufficient.