search-and-replace in cliquet_docs
The name of the thing formerly known as Cliquet is now Kinto-Core. This
isn't super ideal and another cleanup pass will be necessary on this
documentation to make sure it makes sense.

cliquet_fxa is still a real project so we maintain references to it.

This doesn't touch any of the Sphinx configuration, which we'll have to
merge with the Kinto doc config.