Update and fix tests
The last version of the tests had a small imprecision. When we created
the stored record on setUp using the model, we produced a record
with an incremented `last_modified` that was ahead of the server
timestamp. This tries to fix this by creating the record using the
model and than update it using the resource to make the timestamps
match.

This also tries to simplify a bit the tests setting record_id on
the setup.