use json instead of ujson in tests
to catch bytes instead of str passed to storage.
Fixes #1238.

A new boolean settings `storage_strict_json` is introduced
to configure json instead of ujson in tests but keep ujson
in prod. `strict_json` is True by default in `StorageBase`
constructor, but False by default in `load_from_config`.

strict_json is still not in effect in:
 - test_main.py,
 - tests/core/test_initialization.py
 - tests/plugins/test_{history,quotas}.py#test_a_statsd...