Straighten out the comparison order for Postgres
(Or, you could say we "sorted out" the comparison order...)

This isn't necessarily the best possible sort order, but it's the one
that Postgres supports natively, so let's capitalize on that. We may
have to allow different storage backends to define their own sort
order, but we have a couple tests that rely on sort
order (test_get_all_can_deal_with_none_values and
test_get_all_can_filter_with_numeric_values) so let's just go with
this one.