Straighten out comparison order for the memory backend
This is explicitly the same as the Postgres backend, at least as a
first approximation. The full complete comparison order isn't
implemented, and I expect probably won't ever be implemented as I
think the only logical way to handle different storage backends is to
allow them to enforce their own sort order. However, a couple of tests
rely on this sort order for all sort backends, so enforce it.