Fix test
Pagination needs to handle missing fields correctly, because we handle
NULLs (i.e. MISSINGs) as comparing greater than everything
else. Otherwise, we generate a token for a MISSING and resume the next
page on JSON ``null``s, which skips from the top of the sort to the
bottom.

Because this commit promotes MISSING from the memory backend, it works
correctly there. However, the Postgres backend probably needs to be
updated.