Add exc_info to logger call
These exceptions are caught in the exception view and not logged but
converted into HTTP Conflict responses. That means this is the only
place where they are logged. Other exceptions get logged with the
exc_info argument set, which seems like it produces tracebacks and may
be responsible for stack traces getting reported to Sentry. Try to
bring this logger call in line with the others to see if it makes a
difference.