Fix TOCTOU bug in postgresql create
Doing a get() and then an insert() later risks another transaction
inserting something between the get and the insert. The correct way to
handle this is to just go ahead and try the insert but detect
conflicts. In order to do this, we do a complicated
SELECT-or-INSERT-or-UPDATE based on some examples from
StackOverflow. We upsert to replace any tombstones, but we combine
that with a select to detect any existing non-tombstone records.

Writing a test for this is tricky because it requires two
transactions, one which happens in the middle of the get() method, so
let's not even try.