Don't purge_deleted
Perhaps the reason we're seeing increased CPU load is because the
purge_deleted call, which used to do a table scan over a relatively
small table, is now over the entire records table. This probably
doubles the amount of table scans we perform.

Let's disable this purge_deleted for a while. Tombstones will stay
tombstones for now. This won't be a problem for kintowe because
kintowe never actually generates tombstones (even deleted records are
encrypted as "live" records).