Add collations to the timestamps table
The trigger bump_timestamp does a query against the timestamps table
using the parent_id and collection_id from the records table (which
are now collation C). If the timestamps columns are not of collation
C, the primary key indices could potentially include different
sequences of letters that should be considered equal, or equal
sequences of bytes which are considered different by the collation, so
Postgres does not use the index. This makes all queries against the
timestamps table perform a table scan, which performs very
badly. Because every request of a default_bucket does at least one
create(), this makes every single HTTP request very expensive.

If we update the collation of the columns in the timestamp table to
match the records table, the queries should return to their previous speed.