Flushing should not wipe metadata
Flushing should wipe data, but not damage the schema. However, the
version of the schema is wiped, which makes it hard to figure out what
schema we have. There's some hacky code in the storage schema that
assumes that if the schema version is missing, we must be "up to
date", but honestly we shouldn't ever monkey around with the metadata
table in this way. This becomes increasingly important because I don't
want to duplicate this hack in the permission backend migration code.

Fix a couple tests to stop touching the metadata table and instead use
the records table, which is flushed predictably.