Don't use ID as a unique key for any given query
Because parent_id with wildcards can match different objects that have
the same ID, this old use of records indexed by IDs to get the union
of a bunch of records no longer works. Instead, rely on the fact that
the same record, matched multiple times, is the same object in
memory, and use its memory location as its unique ID. This is kind of
a hack, but I couldn't think of a better solution.

Although this is one obvious place where we rely on every object in a
given "result set" having a unique ID, there may be others hidden
elsewhere.

Refs: #1616