Simplify Postgres get_all query
This DISTINCT id leads to difficulties when, as with wildcard
parent_ids, multiple records have the same ID. This manifests as
re-selecting the same record in multiple "pages" because its id
matched once.

This DISTINCT also costs us performance because we have to stick every
record in a hash in order to deduplicate it, and then do an expensive
join afterwards to get the record back. None of this is really
necessary, so just delete it.