Sort before you paginate
Leaving the sorting criteria at the end of the query causes Postgres
to fetch all the records first, count them, and then sort them (which
is probably O(n log n)) so that it can select the right page.

Moving the sorting criteria closest to the data means that Postgres
will try to use an index (if possible/available) to generate records
in the right order. This eliminates the cost of doing a sort. However,
it more or less enforces that we will be scanning records via the
index, even if scanning the whole table would be faster and generate
comparatively few records (i.e. few enough that sorting would be
fast).