Extract EventCollector
Because we handle resource events twice -- once to emit
ResourceChanged and ResourceRead events, and then a second time to
emit AfterResourceChanged and ResourceRead events, we have to keep
track of the drained events. Because we might need to "merge" events
here as well as in the original notify_resource_event function,
extract the logic for managing collections of events into a separate
class, EventCollector. Draining the EventCollector is done with an
iterator that is built around the builtin dict.popitem() method.

This restores the `get_resource_events` API from before this branch
started, but it is now a generator rather than a simple accessor.