Only native_value things from the environment
The `value` should be understood to be a Python value already. If it
wasn't a Python value, it should be deserialized towards the "edges",
not here.

Because we call `from_env` multiple times, this prevents us from
multiply-deserializing values.

Settings loaded from the settings files are still strings, so
deserialize them in `load_default_settings`.