Replace ujson with simdjson for PostgreSQL storage backend, json for Memory storage backend
Resolve #2736 by replacing ujson which cannot serialize or deserialize numbers > 64 bits.

simdjson, a performant json library is used for the PostgreSQL backend while native json is used for the memory storage backend in order to ensure that the unit test demonstrating the issue passes.

Note 1: there is no unit test which verifies that this works with PostgreSQL - I manually modified the test configuration setup however to confirm that it does in fact pass.