Use rapidjson instead of ujson
Replaced pysimdjson with python-rapidjson.

As pysimdjson uses stdlib json to serialize json, serialization performance is subpar compared to ujson. With rapidjson, and the selected options used, performance for serialization should be comparable to that of ujson and simdjson, whereas deserialization is slower than that of ujson but better than that of ujson, according to benchmarks found here:
https://python-rapidjson.readthedocs.io/en/latest/benchmarks.html

For serialization, we do not employ the "number_mode=NM_NATIVE" setting which improves performance, in favour of correctness and in order to resolve #2677.