Update import path for authz consts and interface
Running the test suite reported several warnings about the import path of
certain constants, e.g.:
  Authenticated: "pyramid.security.Authenticated" is deprecated in Pyramid 2.0.
  Adjust your import to "pyramid.authorization.Authenticated"

This commit makes those changes to suppress the warnings that we can control,
as some of these warnings are coming from 3rd party libraries.

This commit  also changes how we import IAuthorizationPolicy. I noticed that we
often imported it from "pyramid.security.Authorization", but it seems that it's
meant to be imported from "pyramid.interfaces", so that's where we import it
from now.